1. Date类型的数据，存储实际上为Int类型，转换为long类型存储。

+++ streaming-api/src/main/java/com/tree/finance/bigdata/hive/streaming/mutation/inspector/LogicalDateObjectInspector.java
@@ -9,7 +9,7 @@ import java.sql.Date;
 
 /**
  * @author Zhengsj
- * Description: Date is logical type in connect avro, it is stored as int value
+ * Description: Date is logical type in connect avro, it is stored as long value
  * Created in 2018/8/22 10:08
  */
 public class LogicalDateObjectInspector extends AbstractPrimitiveJavaObjectInspector implements DateObjectInspector {
@@ -20,19 +20,19 @@ public class LogicalDateObjectInspector extends AbstractPrimitiveJavaObjectInspe
 
     @Override
     public DateWritable getPrimitiveWritableObject(Object o) {
-        return new DateWritable(new Date(((int) o) * 86400000L));
+        return new DateWritable(new Date(((long) o) * 86400000));
     }
 
     public Date get(Object o) {
-        return new Date(((int) o) * 86400000L);
+        return new Date(((long) o) * 86400000);
     }
 
     public Date getPrimitiveJavaObject(Object o) {
-        return o == null ? null : new Date(((int) o) * 86400000L);
+        return o == null ? null : new Date(((long) o) * 86400000L);
     }
 
     public Date create(Object o) {
-        return new Date(((int) o) * 86400000L);
+        return new Date(((long) o) * 86400000L);
     }
 
 }
